# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Surveyvor App - Technical Manual",
        "version": "1.0"
    },
    "paths": {
...
        "/api/invitation/save": {
            "post": {
                "tags": [
                    "Invitation"
                ],
                "summary": "Create an Invitation",
                "description": "Returns the generated Invitation",
                "operationId": "saveInvitation",
                "requestBody": {
                    "description": "Pass user details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "sender_id": {
                                        "type": "number",
                                        "example": "1"
                                    },
                                    "receiver_id": {
                                        "type": "number",
                                        "example": "1"
                                    },
                                    "date_sent": {
                                        "type": "string",
                                        "example": "17/05/2022"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "question": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "invitation_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "sender_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "receiver_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "date_sent": {
                                                        "type": "string",
                                                        "example": "17/05/2022"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
        "/api/teamMember/mine": {
            "get": {
                "tags": [
                    "Team Member"
                ],
                "summary": "Get all the team members related to you",
                "description": "Returns a list of the team members related to you",
                "operationId": "getTeamMembers",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "teamMember": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "user_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "first_name": {
                                                        "type": "string",
                                                        "example": "Dereck"
                                                    },
                                                    "family_name": {
                                                        "type": "string",
                                                        "example": "Cepeda"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "test@surveyvor.com"
                                                    },
                                                    "password": {
                                                        "type": "string",
                                                        "example": "password1234"
                                                    },
                                                    "country_code": {
                                                        "type": "string",
                                                        "example": "ES"
                                                    },
                                                    "avatar": {
                                                        "type": "string",
                                                        "example": "www.myimageprofile.com"
                                                    },
                                                    "telephone": {
                                                        "type": "number",
                                                        "example": "12345678"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/teamMember/all/{team_id}": {
            "get": {
                "tags": [
                    "Team Member"
                ],
                "summary": "Get all the team members by team_id provided",
                "description": "Returns a list of the team members by team_id provided",
                "operationId": "getTeamMembersByTeam",
                "parameters": [
                    {
                        "name": "team_id",
                        "in": "path",
                        "description": "Search by team_id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "teamMember": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "user_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "first_name": {
                                                        "type": "string",
                                                        "example": "Dereck"
                                                    },
                                                    "family_name": {
                                                        "type": "string",
                                                        "example": "Cepeda"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "test@surveyvor.com"
                                                    },
                                                    "password": {
                                                        "type": "string",
                                                        "example": "password1234"
                                                    },
                                                    "country_code": {
                                                        "type": "string",
                                                        "example": "ES"
                                                    },
                                                    "avatar": {
                                                        "type": "string",
                                                        "example": "www.myimageprofile.com"
                                                    },
                                                    "telephone": {
                                                        "type": "number",
                                                        "example": "12345678"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/teamMember/save": {
            "post": {
                "tags": [
                    "Team Member"
                ],
                "summary": "Create a new Team Member",
                "description": "Returns the generated Team Member",
                "operationId": "saveTeamMember",
                "requestBody": {
                    "description": "Pass Team Member details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "team_id": {
                                        "type": "number",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "teamMember": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "team_member_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "team_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "user_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "date_registered": {
                                                        "type": "string",
                                                        "example": "17/05/2022"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/teamMember/saveByInvitation": {
            "post": {
                "tags": [
                    "Team Member"
                ],
                "summary": "Create a new Team Member by invitation link",
                "description": "Returns the generated Team Member",
                "operationId": "saveByInvitation",
                "requestBody": {
                    "description": "Pass Invitation details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "team_id": {
                                        "type": "number",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "teamMember": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "team_member_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "team_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "user_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    },
                                                    "date_registered": {
                                                        "type": "string",
                                                        "example": "17/05/2022"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Data Passed Error"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/teamMember/delete": {
            "post": {
                "tags": [
                    "Team Member"
                ],
                "summary": "Delete a Team Member by Team Id and User (by token provided)",
                "operationId": "deleteByTeam",
                "requestBody": {
                    "description": "Pass Team details to leave",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "team_id": {
                                        "type": "number",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Data Passed Error"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
...
    "security": [
        []
    ]
}